added table header to home list and created confirmation on deleting sensor

+++ ICTLAB/ICTLAB/app/controllers/sensorController.js
@@ -1,23 +1,25 @@
 ﻿angular.module("app").controller("sensorController", ["$scope", "$window", "apiService", "$location",
-    function($scope, $window, apiService, $location) {
+    function ($scope, $window, apiService, $location) {
         $scope.home = $window.home;
         //GET SENSOR LIST
         $scope.sensors = {};
 
-        $scope.removeSensor = function(sensor) {
-            apiService.deleteSensor(sensor).then(
-                function successCallback(result) {
-                    $scope.refreshSensors();
-                    //alert("deleted " + sensor.Name);
-                },
-                function errorCallback(result) {
-                    var x = result;
-                    alert("unable to delete: " + sensor.Name);
-                }
-            );
+        $scope.removeSensor = function (sensor) {
+            if ($window.confirm("Are you sure you want to delete: " + sensor.Name)) {
+                apiService.deleteSensor(sensor).then(
+                    function successCallback(result) {
+                        $scope.refreshSensors();
+                        //alert("deleted " + sensor.Name);
+                    },
+                    function errorCallback(result) {
+                        var x = result;
+                        alert("unable to delete: " + sensor.Name);
+                    }
+                );
+            }
         }
 
-        $scope.refreshSensors = function() {
+        $scope.refreshSensors = function () {
             apiService.getSensorsWithoutReadings($scope.home)
                 .then(
                     function successCallback(result) {
@@ -45,21 +47,21 @@
             home: $scope.home
         }
 
-        $scope.changedFloor = function() {
+        $scope.changedFloor = function () {
             switch ($scope.sensor.floor) {
-            case 0:
-                $scope.sensor.room = $scope.floor0[0];
-                break;
-            case 1:
-                $scope.sensor.room = $scope.floor1[0];
-                break;
-            case 2:
-                $scope.sensor.room = $scope.floor2[0];
-                break;
-            default:
+                case 0:
+                    $scope.sensor.room = $scope.floor0[0];
+                    break;
+                case 1:
+                    $scope.sensor.room = $scope.floor1[0];
+                    break;
+                case 2:
+                    $scope.sensor.room = $scope.floor2[0];
+                    break;
+                default:
             }
         }
-        $scope.addSensor = function() {
+        $scope.addSensor = function () {
             if ($scope.sensor.name !== "" && $scope.sensor.type !== "" && $scope.sensor.targetApiLink !== "" && $scope.sensor.unit !== "" && $scope.sensor.floor !== "" && $scope.sensor.room !== "") {
                 apiService.addSensor($scope.sensor).then(
                     function successCallback(result) {
@@ -82,7 +84,7 @@
         }
 
         // MANAGE SENSORS
-        $scope.toggleSensor = function(sensor) {
+        $scope.toggleSensor = function (sensor) {
             if (sensor.IsActive) {
                 sensor.IsActive = false;
 
@@ -105,21 +107,21 @@
         $scope.detailSensor = {
             id: $window.sensorId,
             home: $scope.home
-    };
+        };
 
-        $scope.getSensorById = function() {
+        $scope.getSensorById = function () {
             apiService.getSensorById($scope.detailSensor).then(
                 function successcallback(result) {
                     $scope.sensorDetailModel = result.data;
                 },
                 function errorCallback(result) {
                     var x = result;
-                    alert(result.statusText);    
+                    alert(result.statusText);
                 });
         }
 
         // UPDATE SENSOR
-        $scope.updateSensor = function() {
+        $scope.updateSensor = function () {
             apiService.updateSensor($scope.sensorDetailModel).then(
                 function successCallback(result) {
                     alert("Successfully updatet " + $scope.sensorDetailModel.Name);
